+++ Userland/Libraries/LibGfx/Font/Font.h
@@ -126,8 +126,8 @@ public:
     virtual Glyph glyph(u32 code_point) const = 0;
     virtual bool contains_glyph(u32 code_point) const = 0;
 
-    virtual u8 glyph_width(u32 code_point) const = 0;
-    virtual int glyph_or_emoji_width(u32 code_point) const = 0;
+    virtual float glyph_width(u32 code_point) const = 0;
+    virtual float glyph_or_emoji_width(u32 code_point) const = 0;
     virtual float glyphs_horizontal_kerning(u32 left_code_point, u32 right_code_point) const = 0;
     virtual u8 glyph_height() const = 0;
     virtual int x_height() const = 0;
