+++ Userland/Libraries/LibGfx/Font/ScaledFont.h
@@ -43,8 +43,8 @@ public:
     virtual u16 weight() const override { return m_font->weight(); }
     virtual Gfx::Glyph glyph(u32 code_point) const override;
     virtual bool contains_glyph(u32 code_point) const override { return m_font->glyph_id_for_code_point(code_point) > 0; }
-    virtual u8 glyph_width(u32 code_point) const override;
-    virtual int glyph_or_emoji_width(u32 code_point) const override;
+    virtual float glyph_width(u32 code_point) const override;
+    virtual float glyph_or_emoji_width(u32 code_point) const override;
     virtual float glyphs_horizontal_kerning(u32 left_code_point, u32 right_code_point) const override;
     virtual int preferred_line_height() const override { return metrics().height() + metrics().line_gap; }
     virtual u8 glyph_height() const override { return m_point_height; }
